requirements.txt
anyio==3.6.1
cassandra-driver==3.25.0
click==8.1.3
cql==1.4.0
fastapi==0.85.0
geomet==0.3.0
idna==3.4
pycodestyle==2.9.1
pydantic==1.10.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
thrift==0.16.0
toml==0.10.2
typing_extensions==4.3.0
